Issue-159 - Move http2 test server to its own file
